Added benchmark for image classification
Added a benchmark for image classification which uses resnet50 and the
small flowers dataset.

In adding the benchmark found that ModelSavePath was not wired, and due
to how benchmarks run I needed this, so wired that up as well.

Additionally removed the sample for resnet50 as it added no value
